refactor del modalUser
