+++ src/components/Header/header.module.scss
@@ -29,10 +29,11 @@
   &NavUser {
     display: flex;
     align-items: center;
+    width: 50%;
     justify-content: space-around;
 
     & > li {
-      width: 100px;
+      width: 120px;
     }
   }
 
@@ -47,19 +48,17 @@
   }
 
   &TextButton {
-    width: 55px;
+    width: 120px;
     height: 20px;
-    font-size: 10px;
     font-style: initial;
     font-weight: bold;
-    padding-top: 2px;
   }
 }
 
 .MuiButtonBase {
-  width: 100px;
   background: #000;
-  font-size: 10px;
+  width: 110px;
+  font-size: 14px;
   padding: 0;
 
   & > span:first-child {
