--- src/components/Modal/ModalUser/index.ts
@@ -1,1 +1,0 @@
-export { default } from './modalUpdateUser';
