+++ src/pages/SignInPage/signInPage.tsx
@@ -1,28 +1,20 @@
-import { useEffect } from 'react';
 import { useNavigate } from 'react-router-dom';
 
-import { useAppDispatch, useAppSelector } from 'redux/hooks';
+import { useAppDispatch } from 'redux/hooks';
 import { useSignInMutation } from 'services';
 import { setCredentials } from 'redux/authSlice';
 
+import Form from 'components/Forms/FormAuthorization';
 import Toast from 'components/Toast/toast';
 
 import { AuthInfoType, ErrorResponse } from 'types';
 
 import styles from 'global-styles/authorization.module.scss';
-import Form from 'components/Forms/FormAuthorization';
 
 const SignInPage = () => {
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
   const [signIn, { error }] = useSignInMutation();
-  const isAuthorized = useAppSelector((state) => state.auth.isAuthorized);
-
-  useEffect(() => {
-    if (isAuthorized) {
-      navigate('/', { replace: true });
-    }
-  });
 
   const onSubmit = async (dataLogin: AuthInfoType) => {
     await signIn(dataLogin)
@@ -30,10 +22,8 @@ const SignInPage = () => {
       .then((data) => {
         localStorage.setItem('pma_token', data.token);
         dispatch(setCredentials(data));
-        setTimeout(() => {
-          navigate('/boards');
-        }, 50);
       })
+      .then(() => navigate('/boards'))
       .catch(() => {
         localStorage.removeItem('pma_token');
       });
