+++ src/pages/UserPage/userPage.utils.ts
@@ -1,0 +1,10 @@
+import { UserSignUpType } from 'types';
+
+export const updateProfile = (id: string, userData: UserSignUpType) => {
+  return {
+    _id: id,
+    name: userData.name,
+    login: userData.login,
+    password: userData.password,
+  };
+};
